goops class objects pointing to c-obj-type
Put this stuff in a musescore scheme file
that will be common for both testing and running.

For testing there is no issue of loading external scheme files.
But for running, we must rely on files and paths which is
installation dependent.

Ie, when launching musescore, it will initialize guile.
Which in turn will try to find and load "musescore.scm" file.