ENG-11: Articulations initially drawn incorrectly
Specifically when dragging-and-dropping an XML file into MuseScore, the
articulations didn't layout correctly. This was due to a difference
between openScore (called when opening from recent or files) and
dropEvent (called when dropping).

This commit removes the partially-duplicated code in dropEvent and
rather calls openScore to handle loading, preventing discrepancies
between the two approaches.

Duplicate of #8285