Add further conditions for tempo inference
This commit adds another case of tempo inference, which handles staff
text on beat 1 of a score (tick 0/1) that is above the staff and bold,
regardless of the presence of keywords.

NOTE: in doing so, it was highlighted in the testing that there is some
existing unreliable behavior with exporting hidden tempo markings. This
test has been temporarily disabled.

Duplicate of #8412, part 4