ENG-9: Limit copyright info to first page
Currently, MuseScore displays copyright metadata in a centered footer.
This results in the information being displayed on every page, which
is inconsistent with many published scores.

This commit hides the footers on import and rather creates a VBox,
populates it with the copyright metadata, and places it at the bottom
of the first page (either below the first explicit pagebreak, or below
the last system of the first page after an initial layout).

Duplicate of #8763, part 1