ENG-58: Fix second-line melisma issue
This commit fixes a relatively rare bug in importing lyrics. Namely,
when a melisma ended on a lyric connected to a note (much less a rest)
on a different voice, the melisma would not be properly ended and extend
until the next opportunity. This changes the conditions by which a
melisma is ended to correct this.

This commit also exports lyrics connected to rests, which previously
were omitted (but sometimes occur, particularly when the lyrics in a
repeat are a different rhythm than the first time).

Duplicate of #8461, part 1