ENG-63: Properly handle cue grace notes
Previously, our importer and exporter didn't account for notes that
were both grace notes and cue notes (therefore extra-small in size).
This commit amends that. Additionally, since Dolet also seems to fail
in this area, this commit adds a function to coerce the `small` and
`play` attributes of a grace chord to match that of its main chord.
Finally, it adds support for the "grace-cue" size to the exporter and
importer. In the importer, this is treated the same as the "cue" size, as
exporters tend to misinterpret the distinction between these.

Duplicate of #8581, plus fixing a compiler warning, see #8554