ENG-66: Repeat text on incorrect side of barline
Sometimes, repeat text (Codas and Segnos) are exported on the incorrect
side of the barline (i.e. end of mm. 29 vs. beginning of mm. 30). Since
MuseScore is unable to place these at a specific tick (and rather attach
them to the right or left side of the measure), this commit adds a more
thorough inference of which measure was originally intended.

Duplicate of #8623